+++ packages/ice/src/createService.ts
@@ -100,6 +100,37 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     plugins.push(pluginWeb());
   }
 
+  // Register framework level API.
+  generatorAPI.addExport({
+    specifier: ['Link', 'Outlet', 'useParams', 'useSearchParams', 'useLocation', 'useNavigate'],
+    source: '@ice/runtime/router',
+  });
+
+  generatorAPI.addExport({
+    specifier: [
+      'defineAppConfig',
+      'useAppData',
+      'useData',
+      'useConfig',
+      'Meta',
+      'Title',
+      'Links',
+      'Scripts',
+      'Data',
+      'Main',
+      'history',
+      'KeepAliveOutlet',
+      'useMounted',
+      'ClientOnly',
+      'withSuspense',
+      'useSuspenseData',
+      'defineDataLoader',
+      'defineServerDataLoader',
+      'defineStaticDataLoader',
+    ],
+    source: '@ice/runtime',
+  });
+
   const ctx = new Context<Config, ExtendsPluginAPI>({
     rootDir,
     command,
@@ -182,7 +213,8 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     hasExportAppData,
     runtimeModules,
     coreEnvKeys,
-    basename: platformTaskConfig.config.basename || '/',
+    // Stringify basename because `config` basename in task config only support type string.
+    basename: JSON.stringify(platformTaskConfig.config.basename || '/'),
     memoryRouter: platformTaskConfig.config.memoryRouter,
     hydrate: !csr,
     importCoreJs: polyfill === 'entry',