+++ packages/ice/CHANGELOG.md
@@ -11,6 +11,7 @@
 - [fix] rebuild server entry when document changed
 - [fix] set ssr default value to `false`
 - [fix] server compiler options when bundle server entry
+- [fix] support define expression of basename
 
 ## v3.0.5